+++ Disciple/src/utopia/disciple/apache/Gateway.scala
@@ -1,31 +1,28 @@
 package utopia.disciple.apache
 
-import org.apache.http.{Consts, HttpEntity}
-import org.apache.http.client.config.RequestConfig
-import org.apache.http.client.entity.UrlEncodedFormEntity
-import org.apache.http.client.methods.{CloseableHttpResponse, HttpDelete, HttpGet, HttpPatch, HttpPost, HttpPut}
-import org.apache.http.client.utils.URIBuilder
-import org.apache.http.impl.client.{HttpClientBuilder, HttpClients}
-import org.apache.http.impl.conn.PoolingHttpClientConnectionManager
-import org.apache.http.message.{BasicHeader, BasicNameValuePair}
-import org.apache.http.ssl.{SSLContextBuilder, SSLContexts}
+import org.apache.hc.client5.http.classic.methods.{HttpDelete, HttpGet, HttpPatch, HttpPost, HttpPut}
+import org.apache.hc.client5.http.config.RequestConfig
+import org.apache.hc.client5.http.entity.UrlEncodedFormEntity
+import org.apache.hc.client5.http.impl.classic.{CloseableHttpResponse, HttpClientBuilder, HttpClients}
+import org.apache.hc.client5.http.impl.io.PoolingHttpClientConnectionManager
+import org.apache.hc.core5.http.{Header, HttpEntity}
+import org.apache.hc.core5.http.message.BasicNameValuePair
+import org.apache.hc.core5.net.URIBuilder
 import utopia.access.http.Method._
 import utopia.access.http.{Headers, Method, Status}
 import utopia.disciple.http.request.TimeoutType.{ConnectionTimeout, ManagerTimeout, ReadTimeout}
 import utopia.disciple.http.request.{Body, Request, Timeout}
 import utopia.disciple.http.response.{ResponseParser, StreamedResponse}
-import utopia.disciple.model.KeyStoreSettings
 import utopia.flow.datastructure.immutable.{Model, Value}
 import utopia.flow.parse.{JSONReader, JsonParser}
 import utopia.flow.time.TimeExtensions._
 import utopia.flow.util.AutoClose._
-import utopia.flow.util.CollectionExtensions._
-import utopia.flow.util.FileExtensions._
 
 import java.io.OutputStream
 import java.net.URLEncoder
-import java.security.KeyStore
-import javax.net.ssl.SSLContext
+import java.nio.charset.StandardCharsets
+import java.util
+import java.util.concurrent.TimeUnit
 import scala.concurrent.{ExecutionContext, Future}
 import scala.io.Codec
 import scala.jdk.CollectionConverters._
@@ -49,9 +46,6 @@ object Gateway
 	  *                          None if no encoding should be used (default).
 	  * @param defaultResponseEncoding Default character encoding used when parsing response data
 	  *                                (used when no character encoding is specified in response headers) (default = UTF-8)
-	  * @param keyStoreSettings Settings to use when interacting with the keystore (SSL certificates) (optional)
-	  * @param trustStoreSettings Settings to use when interacting with the truststore (SSL certificates) (optional).
-	  *                           NB: Only path, password and 'throwErrors' are used.
 	  * @param allowBodyParameters Whether parameters could be moved to request body when body is omitted (default = true).
 	  *                            Use false if you wish to force parameters to uri parameters.
 	  * @param allowJsonInUriParameters Whether uri parameters should be allowed to be converted to json values before
@@ -64,11 +58,9 @@ object Gateway
 	          maxConnectionsTotal: Int = 10,
 	          maximumTimeout: Timeout = Timeout(connection = 5.minutes, read = 5.minutes),
 	          parameterEncoding: Option[Codec] = None, defaultResponseEncoding: Codec = Codec.UTF8,
-	          keyStoreSettings: Option[KeyStoreSettings] = None, trustStoreSettings: Option[KeyStoreSettings] = None,
 	          allowBodyParameters: Boolean = true, allowJsonInUriParameters: Boolean = true) =
 		new Gateway(jsonParsers, maxConnectionsPerRoute, maxConnectionsTotal, maximumTimeout, parameterEncoding,
-			defaultResponseEncoding, keyStoreSettings, trustStoreSettings, b => b,
-			allowBodyParameters, allowJsonInUriParameters)
+			defaultResponseEncoding, b => b, allowBodyParameters, allowJsonInUriParameters)
 	
 	/**
 	  * Creates a new gateway instance
@@ -85,9 +77,6 @@ object Gateway
 	  *                          None if no encoding should be used (default).
 	  * @param defaultResponseEncoding Default character encoding used when parsing response data
 	  *                                (used when no character encoding is specified in response headers) (default = UTF-8)
-	  * @param keyStoreSettings Settings to use when interacting with the keystore (SSL certificates) (optional)
-	  * @param trustStoreSettings Settings to use when interacting with the truststore (SSL certificates) (optional).
-	  *                           NB: Only path, password and 'throwErrors' are used.
 	  * @param allowBodyParameters Whether parameters could be moved to request body when body is omitted (default = true).
 	  *                            Use false if you wish to force parameters to uri parameters.
 	  * @param allowJsonInUriParameters Whether uri parameters should be allowed to be converted to json values before
@@ -101,11 +90,10 @@ object Gateway
 	           maxConnectionsTotal: Int = 10,
 	           maximumTimeout: Timeout = Timeout(connection = 5.minutes, read = 5.minutes),
 	           parameterEncoding: Option[Codec] = None, defaultResponseEncoding: Codec = Codec.UTF8,
-	           keyStoreSettings: Option[KeyStoreSettings] = None, trustStoreSettings: Option[KeyStoreSettings] = None,
 	           allowBodyParameters: Boolean = true, allowJsonInUriParameters: Boolean = true)
 	          (customizeClient: HttpClientBuilder => HttpClientBuilder) =
 		new Gateway(jsonParsers, maxConnectionsPerRoute, maxConnectionsTotal, maximumTimeout, parameterEncoding,
-			defaultResponseEncoding, keyStoreSettings, trustStoreSettings, customizeClient, allowBodyParameters,
+			defaultResponseEncoding, customizeClient, allowBodyParameters,
 			allowJsonInUriParameters)
 }
 
@@ -127,9 +115,6 @@ object Gateway
   *                          None if no encoding should be used (default).
   * @param defaultResponseEncoding Default character encoding used when parsing response data
   *                                (used when no character encoding is specified in response headers) (default = UTF-8)
-  * @param keyStoreSettings Settings to use when interacting with the keystore (SSL certificates) (optional)
-  * @param trustStoreSettings Settings to use when interacting with the truststore (SSL certificates) (optional).
-  *                           NB: Only path, password and 'throwErrors' are used.
   * @param customizeClient A function for customizing the http client when it is first created (default = identity)
   * @param allowBodyParameters Whether parameters could be moved to request body when body is omitted (default = true).
   *                            Use false if you wish to force parameters to uri parameters.
@@ -138,12 +123,10 @@ object Gateway
   *                                 (using .toString). This mostly affects string values, whether they should be
   *                                 wrapped in quotation marks or not. (default = true = use json value format)
 **/
-// FIXME: SSL stuff simply doesn't work (doesn't affect anything)
 class Gateway(jsonParsers: Vector[JsonParser] = Vector(JSONReader), maxConnectionsPerRoute: Int = 2,
               maxConnectionsTotal: Int = 10,
               maximumTimeout: Timeout = Timeout(connection = 5.minutes, read = 5.minutes),
               parameterEncoding: Option[Codec] = None, defaultResponseEncoding: Codec = Codec.UTF8,
-              keyStoreSettings: Option[KeyStoreSettings] = None, trustStoreSettings: Option[KeyStoreSettings] = None,
               customizeClient: HttpClientBuilder => HttpClientBuilder = b => b,
               allowBodyParameters: Boolean = true, allowJsonInUriParameters: Boolean = true)
 {
@@ -160,50 +143,8 @@ class Gateway(jsonParsers: Vector[JsonParser] = Vector(JSONReader), maxConnectio
 	connectionManager.setMaxTotal(maxConnectionsTotal)
 	
 	// TODO: Add customizable timeouts (see https://www.baeldung.com/httpclient-timeout)
-    private lazy val client = {
-	    def _loadTrustStore(base: SSLContextBuilder, settings: KeyStoreSettings) =
-		    settings.storePassword match {
-			    case Some(password) =>
-				    base.loadTrustMaterial(settings.storePath.toFile, password.toCharArray)
-			    case None => base.loadTrustMaterial(settings.storePath.toFile)
-		    }
-	    def _contextToOption(context: Try[SSLContext], throwErrors: Boolean) = {
-		    context.failure.foreach { error => if (throwErrors) throw error else error.printStackTrace() }
-		    context.toOption
-	    }
-	    // Configures the key store / trust store, if necessary
-	    val sslContext = keyStoreSettings match {
-		    // Case KeyStore is used
-		    case Some(settings) =>
-			    val attempt = settings.storePath.readWith { fileStream =>
-				    // Loads the keystore
-				    val store = KeyStore.getInstance(settings.storeType.getOrElse { KeyStore.getDefaultType })
-				    store.load(fileStream, settings.storePassword.map { _.toCharArray }.orNull)
-				    // Creates the SSL context
-				    val base = SSLContexts.custom()
-					    .loadKeyMaterial(store, settings.keyPassword.map { _.toCharArray }.orNull)
-				    // May add the trust store, also
-				    trustStoreSettings.map { _loadTrustStore(base, _) }.getOrElse(base).build()
-			    }
-			    // Handles the failure according to settings
-			    _contextToOption(attempt, settings.throwErrors || trustStoreSettings.exists { _.throwErrors })
-		    // Case: No key store is used => checks trust store settings, still
-		    case None =>
-			    trustStoreSettings.flatMap { settings =>
-				    val attempt = Try { _loadTrustStore(SSLContexts.custom(), settings).build() }
-				    _contextToOption(attempt, settings.throwErrors)
-			    }
-	    }
-	    
-	    val baseClient = HttpClients.custom().setConnectionManager(connectionManager).setConnectionManagerShared(true)
-	    val sslConfiguredClient = sslContext match {
-		    case Some(context) => baseClient.setSSLContext(context)
-		    case None => baseClient
-	    }
-	    // Adds customization according to construction parameters
-	    customizeClient(sslConfiguredClient).build()
-    }
-    
+    private lazy val client = customizeClient(
+	    HttpClients.custom().setConnectionManager(connectionManager).setConnectionManagerShared(true)).build()
     
     // OTHER METHODS    ----------------------
     
@@ -232,14 +173,15 @@ class Gateway(jsonParsers: Vector[JsonParser] = Vector(JSONReader), maxConnectio
 			val config =
 			{
 				val builder = RequestConfig.custom()
-				(request.timeout min maximumTimeout).thresholds.view.mapValues { _.toMillis.toInt }.foreach { case (timeoutType, millis) =>
-					timeoutType match
-					{
-						case ConnectionTimeout => builder.setConnectTimeout(millis)
-						case ReadTimeout => builder.setSocketTimeout(millis)
-						case ManagerTimeout => builder.setConnectionRequestTimeout(millis)
+				(request.timeout min maximumTimeout).thresholds.view.mapValues { _.toMillis.toInt }
+					.foreach { case (timeoutType, millis) =>
+						timeoutType match
+						{
+							case ConnectionTimeout => builder.setConnectTimeout(millis, TimeUnit.MILLISECONDS)
+							case ReadTimeout => builder.setResponseTimeout(millis, TimeUnit.MILLISECONDS)
+							case ManagerTimeout => builder.setConnectionRequestTimeout(millis, TimeUnit.MILLISECONDS)
+						}
 					}
-				}
 				builder.build()
 			}
 			base.setConfig(config)
@@ -393,15 +335,15 @@ class Gateway(jsonParsers: Vector[JsonParser] = Vector(JSONReader), maxConnectio
 	    else
 	    {
 	        val paramsList = params.attributes.map { c => new BasicNameValuePair(c.name, c.value.getString) }
-	        Some(new UrlEncodedFormEntity(paramsList.asJava, Consts.UTF_8))
+	        Some(new UrlEncodedFormEntity(paramsList.asJava, StandardCharsets.UTF_8))
 	    }
 	}
 	
 	private def wrapResponse(response: CloseableHttpResponse) = 
 	{
-	    val status = statusForCode(response.getStatusLine.getStatusCode)
-	    val headers = Headers(response.getAllHeaders.map(h => (h.getName, h.getValue)).toMap)
-	    
+	    val status = statusForCode(response.getCode)
+	    val headers = Headers(response.getHeaders.map(h => (h.getName, h.getValue)).toMap)
+		
 	    new StreamedResponse(status, headers)({ Option(response.getEntity).map { _.getContent } })
 	}
 	
@@ -418,18 +360,32 @@ class Gateway(jsonParsers: Vector[JsonParser] = Vector(JSONReader), maxConnectio
 	{
 		// Inspection is suppressed because this is an implemented function from another library
 		
-		//noinspection ScalaDeprecation
+		/*
 		override def consumeContent() =
 	    {
 	        b.stream.foreach { input =>
     	        val bytes = new Array[Byte](1024) //1024 bytes - Buffer size
                 Iterator.continually (input.read(bytes)).takeWhile (-1 !=).foreach { _ => }
 	        }
-	    }
-	    
-        override def getContent() = b.stream.getOrElse(null)
+	    }*/
+		
+		
+		override def getTrailers = () => new util.ArrayList[Header]()
+		
+		override def close() = {
+			// Consumes and closes the input stream
+			b.stream.foreach { input =>
+				val bytes = new Array[Byte](1024) //1024 bytes - Buffer size
+				Iterator.continually (input.read(bytes)).takeWhile (-1 !=).foreach { _ => }
+				input.close()
+			}
+		}
+		
+		override def getTrailerNames = new util.HashSet[String]()
+		
+		override def getContent() = b.stream.getOrElse(null)
         
-        override def getContentEncoding() = b.contentEncoding.map(new BasicHeader("Content-Encoding", _)).orNull
+        override def getContentEncoding() = b.contentEncoding.orNull
 		
 		override def getContentLength() = b.contentLength.getOrElse(-1)
 		
@@ -443,7 +399,7 @@ class Gateway(jsonParsers: Vector[JsonParser] = Vector(JSONReader), maxConnectio
 				case Some(charset) => s"; charset=${charset.name()}"
 				case None => ""
 			}
-			new BasicHeader("Content-Type", s"${b.contentType}$charsetPart")
+			s"${b.contentType}$charsetPart"
 		}
 		
 		override def isChunked() = b.chunked