+++ Annex/Changes.md
@@ -1,5 +1,8 @@
 # Utopia Annex - List of Changes
 
+## v1.0.4 (in development)
+Dependency changes, according to Disciple update v1.5
+
 ## v1.0.3 - 04.11.2021
 Supports changes in Flow v1.14